Update internal awaitables to use ref counting
There's no way to guarantee the lifetime of an awaitable when it comes in as a CQE. To avoid expensive bookkeeping of active awaitables in the runtime, we instead use intrusive ref counting.

This means a CQE can come in at any time and the relevant awaitable data is still alive, avoiding use-after-free bugs.