Update recv ops to complete with an owned buffer
The codebase is now mostly clean under many of clang's -Weverything warnings.

Multishot recv now uses a new owned buffer abstraction which avoids starvation.

Most of a TLS client has been added as well.

We no longer ignore the `overlays/` directory as we'll be adding boost-buffers to it at some point in the future.